rename : to-auth-route-modal 폴더명 auth-route-modal 로 수정
